ai_service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
google-genai
